requirements.txt
flask
flask-cors
python-telegram-bot>=21.0
pandas
playwright
uvicorn[standard]